+++ src/main/java/io/ejangs/docsa/global/mongoDeleteSystem/app/MongoDeleteRetryService.java
@@ -2,8 +2,8 @@ package io.ejangs.docsa.global.mongoDeleteSystem.app;
 
 import io.ejangs.docsa.domain.block.dao.mongodb.BlockRepository;
 import io.ejangs.docsa.domain.commit.dao.mongodb.CommitBlockSequenceRepository;
-import io.ejangs.docsa.domain.doc.dto.DocDeleteMongoIdsDto;
 import io.ejangs.docsa.domain.save.dao.mongodb.SaveContentRepository;
+import io.ejangs.docsa.global.mongoDeleteSystem.dto.DocDeleteMongoIdsDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.retry.annotation.Backoff;
